Add new assembly file and update workflows and pom.xml
This commit includes the creation of a new 'distribution.xml' file which is used to define the construction and packaging of the project. In the '.github/workflows/azure.yml', steps have been added for ensuring Java 17 setup, checking maven version and deploying to Azure Function with preconfigured secrets. The workflow name has been changed to 'Deploy to Azure Functions' for clarity. In 'pom.xml', the properties section has been updated to use Java version 17. The importance of these modifications lies in their role in preparing the project for packaging and deployment in the Azure Function environment.