"Refactored data object manipulation for improved testing"
This update removed 'Gson' dependency and refactored the message handling syntax in 'EventDtoTest.java' to simplify the testing process. Changes were also implemented in 'EventDto.java', 'Participant.java', and 'EventGridForwardFunction.java' to handle the extraction of necessary information and the checking of event participants. The removal of 'extractParticipant' method from 'Participant', and its incorporation into 'EventDto', simplified the flow and made the codebase more maintainable. An explicit 'Constants' class was added to replace hardcoded strings across the codebase, providing more clarity and ease of modifications in the future. Finally, changes in 'ParticipantTest.java' reflect these modifications.