"Renamed 'filepath' to 'filename' in Participant related code"
We're required to log only the filenames, not the filepath. Therefore, the 'filepath' attribute in Participant related code have been renamed to 'filename'. This change manifests in the Participant data object, the EventGridForwardFunction where the logger now logs the filename, the EventDto where filename is now set for the participant, and relevant tests have been updated to reflect this change. By ensuring the correct naming of the information required to be logged, we add clarity to our system logging.